server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
aiohttp==3.8.6
isodate==0.6.1
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3